requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.4
python-engineio==4.5.1
python-socketio==5.9.0
opencv-python==4.8.1.78
ultralytics==8.0.200
numpy==1.24.3
pillow==10.0.0
eventlet==0.33.3